refactor to fix circular depndency
migrated functions to process_mp4.py
